llmdialect/ir: Usage's cache counts are optional
Usage.CacheReadInputTokens and CacheWriteInputTokens are *int64: nil
when the backend reported no such figure, a pointer to the count, zero
included, when it did, which is how CostUSDMicro already tells unknown
from zero. A plain integer folded "not measured" into a measured zero,
and the Anthropic frontend then told a Messages client that a cache
served nothing where an engine had reported nothing at all.

This is the type change and the mechanical adaptation of every
consumer: the bridge and the frontends dereference with zero for nil,
and each backend still sets the count it decoded, so no wire byte
changes here. Each backend's report is narrowed to what its wire
carried in its own change. The changelog lists the API change under
Changed with the migration.

+++ llmdialect/openaichat/frontend_test.go
@@ -169,7 +169,7 @@ func TestFrontendEncodeResponse(t *testing.T) {
 			{Type: ir.BlockRedactedThinking, Redacted: "xx"},
 		},
 		StopReason: ir.StopToolUse,
-		Usage:      ir.Usage{InputTokens: 40, OutputTokens: 5, CacheReadInputTokens: 60, ReasoningTokens: 2},
+		Usage:      ir.Usage{InputTokens: 40, OutputTokens: 5, CacheReadInputTokens: i64(60), ReasoningTokens: 2},
 	}
 	raw, err := NewFrontend().EncodeResponse(resp)
 	if err != nil {
